refactors server structure, confirms signUp and SignIn work locally
