+++ index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const { db } = require('./lib/models/index');
+const { start } = require('./lib/server');
+const PORT = process.env.PORT || 3000;
+
+db.sync()
+  .then(() => start(PORT))
+  .catch(err => console.log(err));
