+++ lib/models/index.js
@@ -1,0 +1,22 @@
+'use strict';
+
+const { Sequelize, DataTypes } = require('sequelize');
+const DATABASE_URL = process.env.DATABASE_URL || 'sqlite:memory';
+
+const userSchema = require('./user.schema');
+
+const db = new Sequelize(DATABASE_URL, {
+  dialectOptions: {
+    ssl: {
+      require: true,
+      rejectUnauthorized: false,
+    },
+  },
+});
+
+const UserModel = userSchema(db, DataTypes);
+
+module.exports = {
+  db,
+  UserModel,
+};
