+++ package.json
@@ -2,10 +2,10 @@
   "name": "server",
   "version": "1.0.0",
   "description": "Deploys an Express server that implements Basic Authentication, with signup and signin capabilities, using a Postgres database for storage.",
-  "main": "app.js",
+  "main": "index.js",
   "scripts": {
-    "start": "node app.js",
-    "watch": "nodemon app.js",
+    "start": "node index.js",
+    "watch": "nodemon index.js",
     "init:config": "sequelize init:config",
     "db:create": "sequelize db:create",
     "test-watch": "NODE_ENV=test jest --watchAll",
